requirements.txt
fastapi
uvicorn[standard]
openai-agents
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
pytest-mock
httpx
